+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.1.1 (unreleased)
 
+- Added error for `pipelined` and `multi` with block parameter
 - Fixed error with `blpop`, `brpop`, `bzpopmin` and `bzpopmax` with version 4.6+ of `redis` gem
 
 ## 0.1.0 (2020-12-14)